Deletes cinemat.NewSettingsDialog.java since it was an unfinished copy.
cinemat.NewSettingsDialog.java has the same content of
cinemat.SettingsDialog.java and also the constructor name hasn't been
change to the correct new name (e.g. NewSettingsDialog).
In addition there are some lines of codes that refers to Settings.UNIFORME
or similar properties that aren't present in the Settigs class.

This looks like a possible new implementation that has never been
completed and, because of that, it cannot be part of the final
compiled application.