Fixes neg.gif file with bad content.
The java.awt Toolkit loads the file but flags that it encountered
errors during decoding. Probably, this is due to the comment present
in the file added by the software that was used to create it.
The comment was:

This GIF file was assembled with GIF Construction Set from:
Alchemy Mindworks Inc.
P.O. Box 500
Beeton, Ontario
L0G 1A0
CANADA.
This comment block will not appear in files created with a registered version of GIF Construction Set

I have overwritten that file with another one that was in the project
that didn't have the same problem.

This problem was discovered because the application was going into
StackOverflow since the elemag.SimulationDisplay is also registered
as observer when calling the 'drawImage' method. The loop was across
5-6 method calls: SpriteCanvas.paintBackground() -> ... ->
Graphics.drawImage(neg_image, ..., ImageObserver) with 'this' as observer.

The bug is due to the fact that the implemented ImageObserver.imageUpdate()
method doesn't check the 'flags' parameter (bitset), but it just
repaint the image calling super.imageUpdate() method, restarting
the method call loop.